requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=3.15
djangorestframework-simplejwt>=5.2
django-filter>=23.0
psycopg2-binary>=2.9
gunicorn
python-dotenv
whitenoise>=6.5
dj-database-url>=1.0.0
